requirement.txt
numpy==1.26.2
pandas==2.1.4
pefile==2023.2.7
Pillow==10.1.0
pyinstaller==6.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
six==1.16.0
tk==0.1.0
tzdata==2023.3
